refactor: from ProjectDash: ProjectDashVacancies, ProjestDashChatModal, ProjectDashVacancyModal
