+++ src/components/ProjectDashVacancyModal/ProjectDashVacancyModal.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+const ProjectDashVacancyModal = props => {
+  return (
+    <form onSubmit={props.handleSubmitVacancy} name="add-vacancy-form">
+      <label htmlFor="vacancy-title">Role:</label>
+      <input name="vacancy-title" id="vacancy-title" />
+      <label htmlFor="vacancy-description">Description:</label>
+      <input name="vacancy-description" id="vacancy-description" />
+      <label htmlFor="vacancy-skills">Skills:</label>
+      <p>Add a comma after each skill. example: React, CSS</p>
+      <input name="vacancy-skills" id="vacancy-skills" />
+      <button type="submit">Submit</button>
+      <button onClick={props.handleCloseVacancyModal} type="button">
+        Cancel
+      </button>
+    </form>
+  );
+};
+
+export default ProjectDashVacancyModal;
